requirements.txt
python-pptx>=1.0.2
python-docx>=1.1.2
PyYAML>=6.0.2
nbformat>=5.10.4
numpy>=1.26.0
matplotlib>=3.8.0
